+++ .yeoman/templates/containers/src/index.stories.jsx
@@ -3,7 +3,7 @@ import { storiesOf } from '@storybook/react';
 import notes from '../README.md';
 import <%= packageName %> from './index';
 
-storiesOf('Containers|<%= packageName %>', module)
+storiesOf('Containers/<%= packageName %>', module)
   .add(
     'default',
     () => (<<%= packageName %> />),