+++ packages/utilities/psammead-storybook-helpers/src/buildRTLSubstories.test.js
@@ -12,7 +12,7 @@ jest.mock('@storybook/react', () => ({
   })),
   getStorybook: jest.fn(() => [
     {
-      kind: 'Components|Brand',
+      kind: 'Components/Brand',
       fileName: './packages/components/psammead-brand/src/index.stories.jsx',
       stories: [
         {
@@ -24,7 +24,7 @@ jest.mock('@storybook/react', () => ({
       ],
     },
     {
-      kind: 'Components|Caption',
+      kind: 'Components/Caption',
       fileName: './packages/components/psammead-caption/src/index.stories.jsx',
       stories: [
         {
@@ -44,13 +44,13 @@ jest.mock('@storybook/react', () => ({
 afterEach(jest.clearAllMocks);
 
 it('should get all stories', () => {
-  buildRTLSubstories('Components|Brand');
+  buildRTLSubstories('Components/Brand');
 
   expect(getStorybook).toHaveBeenCalled();
 });
 
 it('should add the withServicesKnob decorator so that the default service and service options are configured', () => {
-  buildRTLSubstories('Components|Brand');
+  buildRTLSubstories('Components/Brand');
 
   expect(withServicesKnob.default).toHaveBeenCalledWith({
     defaultService: 'arabic',
@@ -59,21 +59,21 @@ it('should add the withServicesKnob decorator so that the default service and se
 });
 
 it("should build RTL variants of story kind's full suite of stories", () => {
-  buildRTLSubstories('Components|Brand');
+  buildRTLSubstories('Components/Brand');
 
-  expect(storiesOf.mock.calls[0][0]).toEqual('Components|Brand/RTL');
+  expect(storiesOf.mock.calls[0][0]).toEqual('Components/Brand/RTL');
   expect(mockAddStory.mock.calls[0][0]).toEqual('RTL - without brand link');
 
-  expect(storiesOf.mock.calls[1][0]).toEqual('Components|Brand/RTL');
+  expect(storiesOf.mock.calls[1][0]).toEqual('Components/Brand/RTL');
   expect(mockAddStory.mock.calls[1][0]).toEqual('RTL - with brand link');
 
   expect(mockAddStory.mock.calls[2]).toBeUndefined();
 });
 
 it("should build RTL variants of story kind's specified stories", () => {
-  buildRTLSubstories('Components|Brand', { include: ['with brand link'] });
+  buildRTLSubstories('Components/Brand', { include: ['with brand link'] });
 
-  expect(storiesOf.mock.calls[0][0]).toEqual('Components|Brand/RTL');
+  expect(storiesOf.mock.calls[0][0]).toEqual('Components/Brand/RTL');
   expect(mockAddStory.mock.calls[0][0]).toEqual('RTL - with brand link');
 
   expect(storiesOf.mock.calls[1]).toBeUndefined();