+++ packages/components/psammead-radio-schedule/CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- prettier-ignore -->
 | Version | Description |
 |---------|-------------|
+| 3.0.25 | [PR#3831](https://github.com/bbc/psammead/pull/3831) Talos - Bump Dependencies - @bbc/psammead-timestamp-container |
+| 3.0.24 | [PR#3829](https://github.com/bbc/psammead/pull/3829) Talos - Bump Dependencies - @bbc/psammead-timestamp-container |
+| 3.0.23 | [PR#3816](https://github.com/bbc/psammead/pull/3816) Talos - Bump Dependencies |
+| 3.0.22 | [PR#3728](https://github.com/bbc/psammead/pull/3728) Tweaks following Storybook v5.3.19 to v6.0.22 update |
 | 3.0.21 | [PR#3811](https://github.com/bbc/psammead/pull/3811) Talos - Bump Dependencies - @bbc/psammead-styles |
 | 3.0.20 | [PR#3799](https://github.com/bbc/psammead/pull/3799) Talos - Bump Dependencies - @bbc/psammead-timestamp-container |
 | 3.0.19 | [PR#3798](https://github.com/bbc/psammead/pull/3798) Talos - Bump Dependencies